+++ src/containerapp/azext_containerapp/_help.py
@@ -289,6 +289,15 @@ helps['containerapp revision label remove'] = """
           az containerapp revision label remove -n MyContainerapp -g MyResourceGroup --label myLabel
 """
 
+helps['containerapp revision label swap'] = """
+    type: command
+    short-summary: Swap a revision label between two revisions with associated traffic weights.
+    examples:
+    - name: Swap a revision label between two revisions..
+      text: |
+          az containerapp revision label swap -n MyContainerapp -g MyResourceGroup --labels myLabel1 myLabel2
+"""
+
 # Environment Commands
 helps['containerapp env'] = """
     type: group