+++ src/containerapp/azext_containerapp/_utils.py
@@ -857,6 +857,19 @@ def _update_revision_weights(containerapp_def, list_weights):
     return old_weight_sum
 
 
+def _validate_revision_name(cmd, revision, resource_group_name, name):
+    if revision.lower() == "latest":
+        return
+    revision_def = None
+    try:
+        revision_def = ContainerAppClient.show_revision(cmd, resource_group_name, name, revision)
+    except:  # pylint: disable=bare-except
+        pass
+
+    if not revision_def:
+        raise ValidationError(f"Revision '{revision}' is not a valid revision name.")
+
+
 def _append_label_weights(containerapp_def, label_weights, revision_weights):
     if "traffic" not in containerapp_def["properties"]["configuration"]["ingress"]:
         containerapp_def["properties"]["configuration"]["ingress"]["traffic"] = []
@@ -864,6 +877,7 @@ def _append_label_weights(containerapp_def, label_weights, revision_weights):
     if not label_weights:
         return
 
+    bad_labels = []
     revision_weight_names = [w.split('=', 1)[0].lower() for w in revision_weights]  # this is to check if we already have that revision weight passed
     for new_weight in label_weights:
         key_val = new_weight.split('=', 1)
@@ -885,7 +899,10 @@ def _append_label_weights(containerapp_def, label_weights, revision_weights):
                 break
 
         if not is_existing:
-            raise ValidationError(f"No label {label} assigned to any traffic weight.")
+            bad_labels.append(label)
+
+    if len(bad_labels) > 0:
+        raise ValidationError(f"No labels '{', '.join(bad_labels)}' assigned to any traffic weight.")
 
 
 def _update_weights(containerapp_def, revision_weights, old_weight_sum):
@@ -893,7 +910,6 @@ def _update_weights(containerapp_def, revision_weights, old_weight_sum):
     new_weight_sum = sum([int(w.split('=', 1)[1]) for w in revision_weights])
     revision_weight_names = [w.split('=', 1)[0].lower() for w in revision_weights]
     divisor = sum([int(w["weight"]) for w in containerapp_def["properties"]["configuration"]["ingress"]["traffic"]]) - new_weight_sum
-    round_up = True
     # if there is no change to be made, don't even try (also can't divide by zero)
     if divisor == 0:
         return
@@ -903,18 +919,17 @@ def _update_weights(containerapp_def, revision_weights, old_weight_sum):
     for existing_weight in containerapp_def["properties"]["configuration"]["ingress"]["traffic"]:
         if "latestRevision" in existing_weight and existing_weight["latestRevision"]:
             if "latest" not in revision_weight_names:
-                existing_weight["weight"], round_up = _round(scale_factor * existing_weight["weight"], round_up)
+                existing_weight["weight"] = round(scale_factor * existing_weight["weight"])
         elif "revisionName" in existing_weight and existing_weight["revisionName"].lower() not in revision_weight_names:
-            existing_weight["weight"], round_up = _round(scale_factor * existing_weight["weight"], round_up)
+            existing_weight["weight"] = round(scale_factor * existing_weight["weight"])
 
-
-# required because what if .5, .5? We need sum to be 100, so can't round up or down both times
-def _round(number, round_up):
-    import math
-    number = round(number, 2)  # required because we are dealing with floats
-    if round_up:
-        return math.ceil(number), not round_up
-    return math.floor(number), not round_up
+    total_sum = sum([int(w["weight"]) for w in containerapp_def["properties"]["configuration"]["ingress"]["traffic"]])
+    index = 0
+    while total_sum < 100:
+        weight = containerapp_def["properties"]["configuration"]["ingress"]["traffic"][index % len(containerapp_def["properties"]["configuration"]["ingress"]["traffic"])]
+        index += 1
+        total_sum += 1
+        weight["weight"] += 1
 
 
 def _validate_traffic_sum(revision_weights):